+++ lib/src/world/world.dart
@@ -19,6 +19,7 @@ class World {
   void process({Duration delta = const Duration(milliseconds: 16)}) {
     for (var system in _systems) {
       system.process(delta);
+      _entityManager.processDeletionQueue();
     }
   }